Added files_api (migrated functions from db_api that are file related)
Archive dialog can now always be opened
